Clean code: Move ServiceIdFactory to support package.
