+++ src/main/java/org/opentripplanner/netex/mapping/NoticeMapper.java
@@ -3,6 +3,7 @@ package org.opentripplanner.netex.mapping;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Notice;
 import org.opentripplanner.model.impl.EntityById;
+import org.opentripplanner.netex.mapping.support.FeedScopedIdFactory;
 
 /**
  * Maps NeTEx notice to OTP notice.
